white balancing, color ranges, etc
