+++ src/Image/image.py
@@ -1,8 +1,8 @@
+from re import I
 import cv2 as cv
 from cv2.typing import MatLike
 from .scanner import image_scaner
-
-#from Image.image_preprocess import pre_process
+import numpy as np
 
 class Image:
     """
@@ -29,14 +29,15 @@ class Image:
         
         self.img_og = image_og  # original image
 
-        # resizing image to fit in screen
+        # resizing image to fit i screen
         self.img_resized: MatLike = self.resize_2_std(image_og, resize_factor)
         
         # scanning the grid in the image
         w, h = self.img_resized.shape[:2] 
-        self.img_scan = self.resize_2_std(image_scaner(self.img_resized), 1, w, w)
-
-        #cv.imwrite(f"img_{self.id}_scan.jpg", self.img_scan)
+        img_scan = self.resize_2_std(image_scaner(self.img_resized), 1, w, w)
+        self.img_scan = self.white_balance(img_scan)   
+    
+        cv.imwrite(f"img_{self.id}_scan.jpg", self.img_scan)
 
         #self.show_steps()
 
@@ -73,5 +74,34 @@ class Image:
         cv.imshow('scanned', self.img_scan)
         cv.waitKey(0)
         cv.destroyAllWindows()
+
+    def white_balance(self, image):
+
+        reference_size=(10, 10)
+        reference_top_left=(82, 100)
+
+        
+
+        # Create the reference 10x10 square for the reference region for white balancing
+        reference_region = image[reference_top_left[1]:reference_top_left[1] + reference_size[1],
+                                reference_top_left[0]:reference_top_left[0] + reference_size[0]]
+
+        # Calculate the mean RGB values of the reference region - image white baseline value
+        mean_reference = np.mean(reference_region, axis=(0, 1))
+
+        # Scaling factors for each channel
+        scale_factors = 255.0 / mean_reference
+
+        # Apply white balancing to the entire image by multiplying the image to the scale factor
+        balanced_image = cv.merge([cv.multiply(image[:, :, i], scale_factors[i]) for i in range(3)])
+
+        # Clip the values to the valid range [0, 255]
+        balanced_image = np.clip(balanced_image, 0, 255).astype(np.uint8)
+
+        cv.rectangle(balanced_image, reference_top_left, (reference_top_left[0] + reference_size[0], reference_top_left[1] + reference_size[1]), (0, 255, 0), 2)
+        
+        return balanced_image
+        
+
